soa-api/src/requirements.txt
Flask_Caching>=1.7.2
Flask_Limiter>=1.0.1
Flask>=1.1.1
flask_restx>=0.2.0
Flask_Cors>=3.0.8
PySocks>=1.5.6
setuptools>=41.1.0
requests==2.23.0
tweepy==3.9.0
vaderSentiment==3.3.2
locationiq==0.0.2
newsapi-python==0.2.6